feat 实现获取咨询记录接口

+++ src/main/java/org/example/psychologicalcounseling/module/user/register/UserRegisterController.java
@@ -25,7 +25,7 @@ public class UserRegisterController {
     }
 
     @GetMapping("/getUserBy/{uid}")
-    public  ResponseEntity<UserDto> getUserByUid(@PathVariable int uid) {
+    public  ResponseEntity<UserDto> getUserByUid(@PathVariable Long uid) {
         UserDto userDto = userRegisterService.findByUid(uid);
         return ResponseEntity.ok(userDto);
     }
